feat: restructure utilities for server-only and client-only imports; add prisma client extension for role badge color
