+++ package.json
@@ -13,6 +13,7 @@
     "@headlessui/react": "^2.2.0",
     "@heroicons/react": "^2.2.0",
     "@prisma/client": "6.5.0",
+    "client-only": "^0.0.1",
     "clsx": "^2.1.1",
     "es-toolkit": "^1.34.1",
     "firebase": "^11.6.0",
@@ -21,6 +22,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.55.0",
+    "server-only": "^0.0.1",
     "sonner": "^2.0.3",
     "swr": "^2.3.3",
     "temporal-polyfill": "^0.3.0",
@@ -44,5 +46,5 @@
     "typescript": "^5",
     "typescript-eslint": "^8.28.0"
   },
-  "packageManager": "pnpm@10.9.0+sha512.0486e394640d3c1fb3c9d43d49cf92879ff74f8516959c235308f5a8f62e2e19528a65cdc2a3058f587cde71eba3d5b56327c8c33a97e4c4051ca48a10ca2d5f"
+  "packageManager": "pnpm@10.10.0+sha512.d615db246fe70f25dcfea6d8d73dee782ce23e2245e3c4f6f888249fb568149318637dca73c2c5c8ef2a4ca0d5657fb9567188bfab47f566d1ee6ce987815c39"
 }
