+++ src/actions/create-random-seed.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { prisma } from "@/utilities";
+import { prisma } from "@/utilities/server-only";
 
 import type { Mission, User } from "@prisma";
