+++ src/actions/create-role.ts
@@ -1,18 +1,21 @@
 "use server";
 
-import { prisma } from "@/utilities";
+import { BadgeColor } from "@/components/base";
+import { prisma } from "@/utilities/server-only";
 
 import type { Mission } from "@prisma";
 
 export async function createRole({
   missionId,
   name,
+  color,
 }: {
   missionId: Mission["id"];
   name: string;
+  color: BadgeColor;
 }) {
   const role = await prisma.role.create({
-    data: { missionId, name },
+    data: { missionId, name, color },
   });
   return role;
 }
