+++ src/actions/get-users.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { prisma } from "@/utilities";
+import { prisma } from "@/utilities/server-only";
 
 export async function getUsers() {
   return await prisma.user.findMany({ where: { deletedAt: null } });
