+++ src/swr/use-user.ts
@@ -4,7 +4,7 @@ import { getAuth, onAuthStateChanged } from "firebase/auth";
 import useSWR, { mutate } from "swr";
 
 import { getUser } from "@/actions/get-user";
-import { firebaseApp, firebaseAuth } from "@/utilities";
+import { firebaseApp, firebaseAuth } from "@/utilities/client-only";
 
 onAuthStateChanged(firebaseAuth, () => mutate(SWR_KEY_ME));
