+++ src/utilities/client-only/firebase.ts
@@ -1,3 +1,5 @@
+import "client-only";
+
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
