+++ src/utilities/client-only/index.ts
@@ -1,0 +1,3 @@
+import "client-only";
+
+export * from "./firebase";
