+++ src/utilities/index.ts
@@ -1,3 +1,1 @@
-export * from "./firebase";
 export * from "./generate-prng";
-export * from "./prisma-client";
