--- src/utilities/prisma-client.ts
@@ -1,3 +1,0 @@
-import { PrismaClient } from "@prisma";
-
-export const prisma = new PrismaClient();
