+++ src/utilities/server-only/index.ts
@@ -1,0 +1,3 @@
+import "server-only";
+
+export * from "./prisma-client";
